Adding a small note
I want to make notes about the game-engine, and eventually try and separate the data-storage from the client, and write a server for it.  The separation between a client and a server would make it easier to analyze, between the graphics on the client-side of the game-demo, and the server-side that stores the data, and finally figure out how to at least attempt to write a working game-engine.